Feature/SP1-16: Comment Feature added (#18)
* SP1-16: Update Campaign endpoint created.

* SP1-16: Global CORS enabled.

* SP-16: Comment feature finished, now you can comment on campaigns.

+++ Backend/src/main/java/com/nocountry/c930/service/impl/CommentServiceImpl.java
@@ -1,22 +1,28 @@
 package com.nocountry.c930.service.impl;
 
 import com.nocountry.c930.dto.CommentDto;
+import com.nocountry.c930.dto.PostCommentDto;
 import com.nocountry.c930.entity.CampaignEntity;
 import com.nocountry.c930.entity.CommentEntity;
+import com.nocountry.c930.entity.RoleEntity;
 import com.nocountry.c930.entity.UserEntity;
+import com.nocountry.c930.enumeration.RoleName;
 import com.nocountry.c930.mapper.CommentMap;
 import com.nocountry.c930.mapper.exception.ParamNotFound;
 import com.nocountry.c930.repository.CampaignRepository;
 import com.nocountry.c930.repository.CommentRepository;
+import com.nocountry.c930.repository.RoleRepository;
 import com.nocountry.c930.repository.UserRepository;
 import com.nocountry.c930.service.ICommentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.util.Set;
+
 @Service
 public class CommentServiceImpl implements ICommentService {
-    
+
     @Autowired
     UserRepository userRepo;
 
@@ -27,20 +33,23 @@ public class CommentServiceImpl implements ICommentService {
     CampaignRepository campaignRepo;
 
     @Autowired
+    RoleRepository roleRepo;
+
+    @Autowired
     CommentMap commentMap;
 
 
-
     @Override
-    public CommentDto createComment(Long id, CommentDto dto) {
+    public CommentDto createComment(Long id, PostCommentDto dto) {
 
         String userEmail = SecurityContextHolder.getContext().getAuthentication().getName();
         UserEntity user = userRepo.findByEmail(userEmail);
 
         CampaignEntity campaign = campaignRepo.findById(id).orElseThrow(
-                ()-> new ParamNotFound("Campaign doesn't exist"));
+                () -> new ParamNotFound("Campaign doesn't exist"));
 
-        CommentEntity comment = commentMap.commentDto2Entity(dto);
+        CommentEntity comment = new CommentEntity();
+        comment.setDescription(dto.getText());
         comment.setUser(user);
         comment.setCampaign(campaign);
 
@@ -71,11 +80,34 @@ public class CommentServiceImpl implements ICommentService {
     }
 
     @Override
-    public void deleteComment(Long id) {
+    public boolean deleteComment(Long idCampaign, Long idComment) {
 
-        CommentEntity comment = commentRepo.findById(id).orElseThrow(
+        String userEmail = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity userSession = userRepo.findByEmail(userEmail);
+        RoleEntity admin = roleRepo.findByName(RoleName.ROLE_ADMIN);
+
+        CampaignEntity campaign = campaignRepo.findById(idCampaign).orElseThrow(
+                () -> new ParamNotFound("Campaign doesn't exist"));
+
+        CommentEntity comment = commentRepo.findById(idComment).orElseThrow(
                 () -> new ParamNotFound("Comment doesn't exist"));
 
-        commentRepo.delete(comment);
+        if (userSession != comment.getUser() && comment.getUser().getRole() != admin) {
+            return false;
+        } else {
+
+            campaign.getCommentsReceived().remove(comment);
+            commentRepo.delete(comment);
+            return true;
+        }
+    }
+
+    @Override
+    public Set<CommentDto> getCampaignComments(Long id) {
+
+        CampaignEntity campaign = campaignRepo.findById(id).orElseThrow(
+                () -> new ParamNotFound("Campaign doesn't exist"));
+
+        return commentMap.commentEntityList2DtoList(campaign.getCommentsReceived());
     }
 }
